+++ changelog.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- BLIDE-323: Ensure syntax errors appear in the editor
+- BLIDE-297: Unscoped variable diagnostics on multiline assignments now highlight only the assignment line
+- BLIDE-280: Workspace scanning now respects root and nested `.gitignore` files
+
 ## [1.11.0] - 2026-05-30
 
 ### Added